requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
pytube==15.0.0
faster-whisper==1.0.1
ffmpeg-python==0.2.0
requests==2.31.0